New Token: 'P' Pulse the laser for a single pulse
takes the power & pulse_duration settings from a M649/G1/G2/G3 command
and fires a single pulse at the current position.

This is intended as debugging aid or for help during alignment of the
laser system. No parameters are to be given, just a single 'P'.

This is similiar to a single 'V' to request the firmware version.

Idea: 'H' may HOME.